+++ ddataflow/data_source.py
@@ -4,6 +4,7 @@ import os
 from pyspark.sql import DataFrame
 
 from ddataflow.exceptions import BiggerThanMaxSize
+from ddataflow.sampling.default import filter_function
 from ddataflow.utils import get_or_create_spark
 
 
@@ -21,11 +22,26 @@ class DataSource:
         snapshot_path: str,
         size_limit,
     ):
-        self.name = name
-        self.config = config
-        self.local_data_folder = local_data_folder
-        self.snapshot_path = snapshot_path
+        self._name = name
+        self._local_data_folder = local_data_folder
+        self._snapshot_path = snapshot_path
         self._size_limit = size_limit
+        self._config = config
+
+        if "source" in self._config:
+            self._source = config["source"]
+        else:
+            if self._config.get('file-type') == 'parquet':
+                self._source = lambda spark: spark.read.parquet(self._name)
+            else:
+                self._source = lambda spark: spark.table(self._name)
+
+        if "filter" in self._config:
+            self._filter = self._config["filter"]
+        else:
+            if self._config.get('default_sampling'):
+                self._filter = lambda df: filter_function(df)
+
 
     def query(self):
         """
@@ -33,11 +49,11 @@ class DataSource:
         """
         df = self.query_without_filter()
 
-        if self.config.get("filter") is not None:
-            print(f"Filter set for {self.name}, applying it")
-            df = self.config["filter"](df)
+        if self._config.get("filter") is not None:
+            print(f"Filter set for {self._name}, applying it")
+            df = self._filter(df)
         else:
-            print(f"No filter set for {self.name}")
+            print(f"No filter set for {self._name}")
 
         return df
 
@@ -46,11 +62,11 @@ class DataSource:
         Go to the raw data source without any filtering
         """
         spark = get_or_create_spark()
-        logger.debug(f"Querying without filter source: '{self.name}'")
-        return self.config["source"](spark)
+        logger.debug(f"Querying without filter source: '{self._name}'")
+        return self._source(spark)
 
     def query_locally(self):
-        logger.info(f"Querying locally {self.name}")
+        logger.info(f"Querying locally {self._name}")
 
         path = self.get_local_path()
         if not os.path.exists(path):
@@ -65,20 +81,20 @@ class DataSource:
         return df
 
     def get_dbfs_sample_path(self) -> str:
-        return os.path.join(self.snapshot_path, self.get_name())
+        return os.path.join(self._snapshot_path, self.get_name())
 
     def get_local_path(self) -> str:
-        return os.path.join(self.local_data_folder, self.get_name())
+        return os.path.join(self._local_data_folder, self.get_name())
 
     def get_name(self) -> str:
-        return self.name
+        return self._name
 
     def get_parquet_filename(self) -> str:
-        return self.name + ".parquet"
+        return self._name + ".parquet"
 
     def estimate_size_and_fail_if_too_big(self):
         """
-        Estimate the size of the data source use the name used in the config
+        Estimate the size of the data source use the _name used in the _config
         It will throw an exception if the estimated size is bigger than the maximum allowed in the configuration
         """
 
@@ -89,13 +105,13 @@ class DataSource:
         print("Estimated size of the Dataset in GB: ", size_estimation)
 
         if size_estimation > self._size_limit:
-            raise BiggerThanMaxSize(self.name, size_estimation, self._size_limit)
+            raise BiggerThanMaxSize(self._name, size_estimation, self._size_limit)
 
         return df
 
-    def _estimate_size(self, df: DataFrame):
+    def _estimate_size(self, df: DataFrame) -> float:
         """
-        Estimatest the size of a dataframe in Gigabytes
+        Estimates the size of a dataframe in Gigabytes
 
         Formula:
             number of gigabytes = (N*V*W) / 1024^3