+++ ddataflow/sampling/default.py
@@ -1,17 +1,18 @@
-from typing import Optional, List
+from datetime import datetime, timedelta
+from typing import List, Optional
+
 import pyspark.sql.functions as F
 from pyspark.sql import DataFrame
 
-from datetime import datetime, timedelta
-
 DEFAULT_SAMPLING_SIZE = 1000
+
 
 def sample_by_yesterday(df: DataFrame) -> DataFrame:
     """
     Sample by yesterday
     """
 
-    yesterday = (datetime.now() - timedelta(days=1)).date();
+    yesterday = (datetime.now() - timedelta(days=1)).date()
 
     if "date" in df.columns:
         print("Found a date column, sampling by yesterday")
@@ -19,7 +20,8 @@ def sample_by_yesterday(df: DataFrame) -> DataFrame:
 
     return df
 
-def filter_funciton(df: DataFrame) -> DataFrame:
+
+def filter_function(df: DataFrame) -> DataFrame:
     """
     Default filter function
     :param df:
@@ -30,7 +32,8 @@ def filter_funciton(df: DataFrame) -> DataFrame:
 
     return df
 
-def build_default_sampling_for_sources(sources: Optional[List[dict]]=None) -> dict:
+
+def build_default_sampling_for_sources(sources: Optional[List[dict]] = None) -> dict:
     """
     Setup standard filters for the entries that we do not specify them
     """
@@ -42,7 +45,7 @@ def build_default_sampling_for_sources(sources: Optional[List[dict]]=None) -> di
         print("Build default sampling for source: " + source)
         result[source] = {
             "source": lambda spark: spark.table(source),
-            "filter": lambda df: filter_funciton(df),
+            "filter": lambda df: filter_function(df),
         }
 
     return result